+++ grails-app/controllers/org/broadinstitute/orsp/UrlMappings.groovy
@@ -96,7 +96,7 @@ class UrlMappings {
         '/api/dur'(controller: 'dataUse', action:'findDataUseRestrictions')
         '/api/collection-links'(controller: 'report', action: 'findCollectionLinks')
         '/api/qa-event-report'(controller: 'statusEvent', action: 'findQaEventReport', method: 'GET')
-        '/api/qa-project-event-report'(controller: 'statusEvent', action:'project', method:'GET')
+        '/api/qa-project-event-report'(controller: 'statusEvent', action: 'findProjectReport', method:'GET')
 
         '/api/comments/save'(controller: 'comments', action: 'saveNewComment', method: 'POST')
         '/api/comments/list'(controller: 'comments', action: 'getComments', method: 'GET')
