+++ src/main/webapp/qaReport/ProjectReport.js
@@ -2,17 +2,17 @@ import React, { Component } from 'react';
 import { div, h, h1, hh } from 'react-hyperscript-helpers';
 import { Link } from 'react-router-dom'
 import { TableComponent } from '../components/TableComponent';
-import { defaultSorted, formatAge, SIZE_PER_PAGE_LIST_PROJECT } from '../util/QaReportConstants';
+import { defaultSorted, formatAge, SIZE_PER_PAGE_LIST_PROJECT, PROJECT_DATA_FIELD } from '../util/ReportsConstants';
 import LoadingWrapper from '../components/LoadingWrapper';
 import { Reports } from '../util/ajax';
 
 const columns = [{
-  dataField: 'event.id',
+  dataField: PROJECT_DATA_FIELD.ID,
   text: 'Id',
   hidden: true,
   csvExport : false
 }, {
-  dataField: 'event.summary',
+  dataField: PROJECT_DATA_FIELD.SUMMARY,
   text: 'Status',
   sort: true,
   headerStyle: (column, colIndex) => {
@@ -21,7 +21,7 @@ const columns = [{
     };
   }
 }, {
-  dataField: 'event.author',
+  dataField: PROJECT_DATA_FIELD.AUTHOR,
   text: 'Author',
   sort: true,
   headerStyle: (column, colIndex) => {
@@ -30,7 +30,7 @@ const columns = [{
     };
   }
 }, {
-  dataField: 'event.created',
+  dataField: PROJECT_DATA_FIELD.CREATED,
   text: 'Status Date',
   sort: true,
   headerStyle: (column, colIndex) => {
@@ -42,7 +42,7 @@ const columns = [{
      new Date(cell).toLocaleString()
 
 }, {
-  dataField: 'duration',
+  dataField: PROJECT_DATA_FIELD.DURATION,
   text: 'Duration',
   sort: true,
   headerStyle: (column, colIndex) => {
